Rename modules per discussion
