+++ src/globus_sdk/experimental/auth_requirements_error/__init__.py
@@ -1,14 +1,14 @@
-from ._validators import ValidationError
-from .auth_requirements_error import (
+from ._auth_requirements_error import (
     GlobusAuthorizationParameters,
     GlobusAuthRequirementsError,
 )
-from .utils import (
+from ._functional_api import (
     has_auth_requirements_errors,
     is_auth_requirements_error,
     to_auth_requirements_error,
     to_auth_requirements_errors,
 )
+from ._validators import ValidationError
 
 __all__ = [
     "ValidationError",
