+++ src/globus_sdk/experimental/auth_requirements_error/_auth_requirements_error.py
